refactor: renamed variables, parameters and devide functions
