+++ src/function/labExamAxios.js
@@ -1,0 +1,11 @@
+import axios from 'axios';
+
+export default function labExamAxios(limit, setData) {
+  axios.get('http://localhost:3001/sample')
+    .then(res => {
+      setData(res.data.slice(0, limit));
+    })
+    .catch((error) => {
+      console.error("Error fetching data:", error);
+    });
+}
